Move preprocessProfileCOntents to apple_mdm package
