+++ ee/server/service/errors.go
@@ -20,42 +20,6 @@ func (e notFoundError) IsNotFound() bool {
 	return true
 }
 
-type NDESInvalidError struct {
-	msg string
-}
-
-func (e NDESInvalidError) Error() string {
-	return e.msg
-}
-
-func NewNDESInvalidError(msg string) NDESInvalidError {
-	return NDESInvalidError{msg: msg}
-}
-
-type NDESPasswordCacheFullError struct {
-	msg string
-}
-
-func (e NDESPasswordCacheFullError) Error() string {
-	return e.msg
-}
-
-func NewNDESPasswordCacheFullError(msg string) NDESPasswordCacheFullError {
-	return NDESPasswordCacheFullError{msg: msg}
-}
-
-type NDESInsufficientPermissionsError struct {
-	msg string
-}
-
-func (e NDESInsufficientPermissionsError) Error() string {
-	return e.msg
-}
-
-func NewNDESInsufficientPermissionsError(msg string) NDESInsufficientPermissionsError {
-	return NDESInsufficientPermissionsError{msg: msg}
-}
-
 type InvalidIDPTokenError struct{}
 
 func (e InvalidIDPTokenError) Error() string {
