+++ ee/server/service/scep/scep_proxy.go
@@ -1,4 +1,4 @@
-package service
+package scep
 
 import (
 	"bytes"
@@ -420,4 +420,40 @@ func (s *SCEPConfigService) GetSmallstepSCEPChallenge(ctx context.Context, ca fl
 	}
 
 	return string(b), nil
+}
+
+type NDESInvalidError struct {
+	msg string
+}
+
+func (e NDESInvalidError) Error() string {
+	return e.msg
+}
+
+func NewNDESInvalidError(msg string) NDESInvalidError {
+	return NDESInvalidError{msg: msg}
+}
+
+type NDESPasswordCacheFullError struct {
+	msg string
+}
+
+func (e NDESPasswordCacheFullError) Error() string {
+	return e.msg
+}
+
+func NewNDESPasswordCacheFullError(msg string) NDESPasswordCacheFullError {
+	return NDESPasswordCacheFullError{msg: msg}
+}
+
+type NDESInsufficientPermissionsError struct {
+	msg string
+}
+
+func (e NDESInsufficientPermissionsError) Error() string {
+	return e.msg
+}
+
+func NewNDESInsufficientPermissionsError(msg string) NDESInsufficientPermissionsError {
+	return NDESInsufficientPermissionsError{msg: msg}
 }
